Move concurrency test to separate class
